refactor Tests
